backend/requirements.txt
# ── AI / LLM ──────────────────────────────────────────────────
anthropic>=0.40.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.50

# ── OpenAI (embeddings only) ───────────────────────────────────
openai>=1.50.0

# ── Vector Store ───────────────────────────────────────────────
chromadb>=0.5.0

# ── Web Search ─────────────────────────────────────────────────
tavily-python>=0.3.0        # Tavily API（可選，無 key 則 fallback DuckDuckGo）

# ── Web Scraping ────────────────────────────────────────────────
httpx>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ── Web Framework ──────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.29.0

# ── Config / Env ───────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# ── Logging ────────────────────────────────────────────────────
loguru>=0.7.0

# ── Async ──────────────────────────────────────────────────────
aiofiles>=23.2.0

# ── Dev ────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.23.0
